Add assetIdForElement API to Resolver
The details of both the `fullName` and `uri` of an `element.source` can vary based on the resolver implementation. Add an API to Resolver so it can own the details without having to leak resolver implementation behavior into `AssetId.resolve` or other APIs.

While all implementations _could_ be synchronous, this is async for the sake of consistency and future proofing.

- Add the new API.
- Implement the API in each Resolver.
- Migrate Angular to use the new API.

This will be submitted internally and, once proven to work, will be published externally.

PiperOrigin-RevId: 225885619